backend/pyproject.toml
[project]
name = "flow-sentinel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "fastapi>=0.128.0",
    "psycopg[binary,pool]>=3.3.2",
    "python-logging-loki>=0.3.1",
    "questdb>=4.1.0",
    "uvicorn>=0.40.0",
]
